Refactor constants Refactor routes Update dev script
