+++ backend/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "whatabyte-portal",
+  "name": "canvas-teams-server",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon ./dist/index.js",
+    "dev": "tsc && concurrently \"tsc -w\" \"nodemon ./dist/index.js\"", 
     "build": "tsc"
   },
   "keywords": [],
@@ -16,6 +16,7 @@
     "@types/multer": "^1.4.5",
     "@types/node": "^14.14.20",
     "@types/passport": "^1.0.5",
+    "concurrently": "^5.3.0",
     "nodemon": "^2.0.4",
     "prettier": "^2.2.1",
     "ts-node": "^9.1.1",
