+++ backend/src/config/constants.ts
@@ -1,0 +1,6 @@
+// @TODO: source appropriate constants from process.env
+export const CLIENT_ID = "10000000000004";
+export const CLIENT_SECRET =
+  "SULCuWVF4E1foAy5GKm7z7hQxWUcOhLZjoBvzSdBKzN3OyNW776oaT1h69O65ozk";
+export const PORT = 3000;
+export const HOST = `http://localhost:${PORT}`
